requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
accelerate=0.27.2=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.4.4=pypi_0
alembic=1.13.2=pypi_0
annotated-types=0.6.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
autoattack=0.1=pypi_0
beautifulsoup4=4.12.3=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
class-resolver=0.4.3=pypi_0
click=8.1.7=pypi_0
click-default-group=1.2.4=pypi_0
colorlog=6.8.2=pypi_0
comm=0.2.1=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
dataclasses-json=0.6.7=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
docdata=0.0.3=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.7.0=pypi_0
entmax=1.3=pypi_0
evdev=1.7.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=h6a678d5_0
filelock=3.13.1=pypi_0
fonttools=4.47.2=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.12.2=pypi_0
gdown=5.1.0=pypi_0
geotorch=0.3.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.62.0=pypi_0
huggingface=0.0.1=pypi_0
huggingface-hub=0.20.3=pypi_0
idna=3.6=pypi_0
imageio=2.34.1=pypi_0
importlib-metadata=7.0.1=pyha770c72_0
importlib_metadata=7.0.1=hd8ed1ab_0
ipykernel=6.29.0=pyhd33586a_0
ipython=8.21.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
joblib=1.4.0=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.5.0=pypi_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.5.0=py312h06a4308_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
mako=1.3.5=pypi_0
markdown=3.5.2=pypi_0
markupsafe=2.1.4=pypi_0
marshmallow=3.21.3=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
more-click=0.1.2=pypi_0
more-itertools=10.3.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.8=pypi_0
munch=4.0.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
numpy=1.26.3=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openssl=3.0.13=h7f8727e_0
optuna=3.6.1=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.2.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=23.3.1=pypi_0
pix2tex=0.1.2=pypi_0
platformdirs=4.2.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=16.1.0=pypi_0
pydantic=2.7.0=pypi_0
pydantic-core=2.18.1=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pykeen=1.10.2=pypi_0
pynput=1.7.6=pypi_0
pyparsing=3.1.1=pypi_0
pyqt6=6.6.1=pypi_0
pyqt6-qt6=6.6.3=pypi_0
pyqt6-sip=13.6.0=pypi_0
pyqt6-webengine=6.6.0=pypi_0
pyqt6-webengine-qt6=6.6.3=pypi_0
pyside6=6.7.0=pypi_0
pyside6-addons=6.7.0=pypi_0
pyside6-essentials=6.7.0=pypi_0
pysocks=1.7.1=pypi_0
pystow=0.5.4=pypi_0
python=3.12.1=h996f2a0_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-xlib=0.33=pypi_0
pytorch-ignite=0.5.0.post2=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
ray=2.31.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
robustbench=1.1=pypi_0
rpds-py=0.18.1=pypi_0
safetensors=0.4.2=pypi_0
scikit-image=0.23.2=pypi_0
scikit-learn=1.4.2=pypi_0
scipy=1.12.0=pypi_0
screeninfo=0.8.1=pypi_0
seaborn=0.13.2=pypi_0
sentry-sdk=1.40.6=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=pypi_0
shiboken6=6.7.0=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.31=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.4.0=pypi_0
tifffile=2024.4.18=pypi_0
timm=0.5.4=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
torch=2.2.0=pypi_0
torch-max-mem=0.1.3=pypi_0
torch-ppr=0.0.8=pypi_0
torchdiffeq=0.2.3=pypi_0
torchvision=0.17.0=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pyhd8ed1ab_0
transformers=4.39.0=pypi_0
triton=2.2.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2024.1=pypi_0
urllib3=2.2.0=pypi_0
wandb=0.16.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=pypi_0
x-transformers=0.15.0=pypi_0
xz=5.4.5=h5eee18b_0
yacs=0.1.8=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
